Update brute_force.js
This was the code snippet that worked.  You need to set the right form submission method which is POST.  Also, I looked for Login Failed instead of the hidden tag, I thought this would not be intuitive and might cause more confusion for folks who are trying to read the code.